add empty check in s3 example
If delete_objects is empty the request will fail XML-validation, I think. Perhaps it's nice to have a check included in the example?

+++ rustv1/examples/s3/src/s3-service-lib.rs
@@ -42,18 +42,20 @@ pub async fn delete_objects(client: &Client, bucket_name: &str) -> Result<Vec<St
     }
 
     let return_keys = delete_objects.iter().map(|o| o.key.clone()).collect();
-
-    client
-        .delete_objects()
-        .bucket(bucket_name)
-        .delete(
-            Delete::builder()
-                .set_objects(Some(delete_objects))
-                .build()
-                .map_err(Error::from)?,
-        )
-        .send()
-        .await?;
+    
+    if !delete_objects.is_empty() {
+        client
+            .delete_objects()
+            .bucket(bucket_name)
+            .delete(
+                Delete::builder()
+                    .set_objects(Some(delete_objects))
+                    .build()
+                    .map_err(Error::from)?,
+            )
+            .send()
+            .await?;
+    }
 
     let objects: ListObjectsV2Output = client.list_objects_v2().bucket(bucket_name).send().await?;
